Refactoring and fine-tuning the footer
