+++ src/utils/global.css
@@ -3,3 +3,7 @@ body {
   padding: 0;
   margin: 0;
 }
+a {
+  text-decoration: none;
+  color: inherit;
+}
